requirements.txt
flask
python-dotenv
minio